Move inputPLC definitions to dedicated header file
